% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM}
\format{
A data frame with 15 rows and 2 columns variables:
\describe{
  \item{DISTANCE}{distance away from target}
  \item{DAMAGE}{HP damage the target took}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
fire
}
\description{
A dataset containing the values for firing distance vs. damage
}
\details{
This is a standard data set featured in the courses of the MSAS program at OU taken provided by the textbook.
}
\keyword{datasets}
