% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmblast.R
\name{lmblast}
\alias{lmblast}
\title{lmblast}
\usage{
lmblast(obj, alpha = 0.05)
}
\arguments{
\item{obj}{use `lm()` to create the obj}

\item{alpha}{the alpha level}
}
\value{
a named list containing the following: R-Squared, a matrix of the residuals, the design and response matrices, the AIC, sample variance, and a confidence interval for Beta's.
}
\description{
This function is exploding with useful regression data produced from the `lm()` function
}
\examples{
\dontrun{lmblast}
}
