% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mygd.R
\name{mygd}
\alias{mygd}
\title{mygd}
\usage{
mygd(binit = c(), s = 1e-05, iter = 500, ylm = ylmf, x, y)
}
\arguments{
\item{binit}{initial estimate of beta's}

\item{s}{learning rate}

\item{iter}{number of iterations}

\item{ylm}{a pre-populated `lm()`}

\item{x}{a numeric vector with the explanatory variable}

\item{y}{a numeric vector with the response variable}
}
\value{
Three plots and an invisible list containing the tail data of the iterated estimates of the Intercept and Slope
}
\description{
A function which produces three plots based on Gradient Descent Theory. The plots estimate, you guessed it, the Beta's
}
\examples{
\dontrun{mygd(binit=c(10,4), s=0.001, iter=500, ylm, x, y)}
}
