% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.R
\docType{data}
\name{spruce}
\alias{spruce}
\title{SPRUCE}
\format{
A data frame with 15 rows and 2 columns variables:
\describe{
  \item{BHDiameter}{breast  height diameter}
  \item{Height}{Height of tree}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
spruce
}
\description{
A dataset containing data about spruce trees.
}
\details{
This is a standard data set featured in the courses of the MSAS program at OU taken provided by the textbook.
}
\keyword{datasets}
